refactor: components module reorganizing
