+++ components/src/commonMain/kotlin/me/lincolnstuart/funblocks/essentials/core/helper/SelectableItem.kt
@@ -1,4 +1,4 @@
-package me.lincolnstuart.funblocks.essentials.helper
+package me.lincolnstuart.funblocks.essentials.core.helper
 
 import androidx.compose.ui.graphics.vector.ImageVector
