+++ components/src/commonMain/kotlin/me/lincolnstuart/funblocks/essentials/core/helper/SimpleItem.kt
@@ -1,4 +1,4 @@
-package me.lincolnstuart.funblocks.essentials.helper
+package me.lincolnstuart.funblocks.essentials.core.helper
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
