+++ components/src/commonMain/kotlin/me/lincolnstuart/funblocks/essentials/misc/cardinfo/CardInfo.kt
@@ -9,10 +9,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.datasource.LoremIpsum
 import me.lincolnstuart.funblocks.essentials.core.spacer.VerticalSpacer
+import me.lincolnstuart.funblocks.essentials.core.surface.Card
 import me.lincolnstuart.funblocks.essentials.core.surface.Surface
 import me.lincolnstuart.funblocks.essentials.misc.cardinfo.utils.CardBodyOptions
 import me.lincolnstuart.funblocks.essentials.misc.cardinfo.utils.CardHeaderOptions
-import me.lincolnstuart.funblocks.essentials.surface.card.Card
 import me.lincolnstuart.funblocks.foundation.ui.theme.FunBlocksTheme
 import me.lincolnstuart.funblocks.foundation.ui.token.content.size.FunBlocksContentSize
 import me.lincolnstuart.funblocks.foundation.ui.token.content.spacing.FunBlocksSpacing
