+++ components/src/commonMain/kotlin/me/lincolnstuart/funblocks/essentials/misc/chip/Chip.kt
@@ -12,11 +12,11 @@ import compose.icons.TablerIcons
 import compose.icons.tablericons.Calendar
 import compose.icons.tablericons.ChevronDown
 import compose.icons.tablericons.X
+import me.lincolnstuart.funblocks.essentials.core.helper.clickable.Clickable
+import me.lincolnstuart.funblocks.essentials.core.helper.clickable.utils.ClickableOptions
 import me.lincolnstuart.funblocks.essentials.core.icon.utils.IconSize
 import me.lincolnstuart.funblocks.essentials.core.spacer.VerticalSpacer
 import me.lincolnstuart.funblocks.essentials.core.surface.Surface
-import me.lincolnstuart.funblocks.essentials.helper.clickable.Clickable
-import me.lincolnstuart.funblocks.essentials.helper.clickable.utils.ClickableOptions
 import me.lincolnstuart.funblocks.essentials.misc.chip.utils.ChipOptions
 import me.lincolnstuart.funblocks.foundation.ui.theme.FunBlocksTheme
 import me.lincolnstuart.funblocks.foundation.ui.token.color.FunBlocksColors
