+++ playground/src/main/java/me/lincolnstuart/funblocks/playground/screens/components/form/select/MultipleSelectScreen.kt
@@ -6,9 +6,9 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import cafe.adriel.voyager.core.screen.Screen
+import me.lincolnstuart.funblocks.essentials.core.helper.SelectableItem
 import me.lincolnstuart.funblocks.essentials.form.select.MultipleSelect
 import me.lincolnstuart.funblocks.essentials.form.switchbutton.SwitchButtonOption
-import me.lincolnstuart.funblocks.essentials.helper.SelectableItem
 import me.lincolnstuart.funblocks.foundation.ui.token.content.spacing.FunBlocksInset
 import me.lincolnstuart.funblocks.playground.components.Sample
