+++ playground/src/main/java/me/lincolnstuart/funblocks/playground/screens/components/misc/ProgressScreen.kt
@@ -9,8 +9,8 @@ import androidx.compose.ui.Modifier
 import cafe.adriel.voyager.core.screen.Screen
 import cafe.adriel.voyager.navigator.LocalNavigator
 import cafe.adriel.voyager.navigator.currentOrThrow
+import me.lincolnstuart.funblocks.essentials.core.list.SimpleListItem
 import me.lincolnstuart.funblocks.essentials.core.surface.Surface
-import me.lincolnstuart.funblocks.essentials.list.SimpleListItem
 import me.lincolnstuart.funblocks.foundation.ui.theme.FunBlocksTheme
 import me.lincolnstuart.funblocks.playground.screens.components.misc.progress.ProgressBarScreen
 import me.lincolnstuart.funblocks.playground.screens.components.misc.progress.StepperScreen
